Update BASE_URL and add startup script for demo
Changed the BASE_URL in App.java to point to the new server at london3.texngine.org, and increased the thread sleep duration to 200ms to reduce polling frequency. Added a new startup script start.sh to streamline the launching process of the tkiz-svg-demo application and redirect output logs to a file.